split netting channel tests
