+++ raiden/tests/smart_contracts/netting_channel/AuxiliaryTester.sol
